meetup page modularize
