+++ app/meetups/hooks/useAnimationVariants.ts
@@ -1,0 +1,30 @@
+import { useMemo } from 'react';
+import { Variants } from 'framer-motion';
+
+export function useAnimationVariants() {
+  const containerVariants: Variants = useMemo(
+    () => ({
+      hidden: {},
+      visible: {
+        transition: {
+          staggerChildren: 0.2,
+        },
+      },
+    }),
+    []
+  );
+
+  const itemVariants: Variants = useMemo(
+    () => ({
+      hidden: { opacity: 0, y: 50 },
+      visible: {
+        opacity: 1,
+        y: 0,
+        transition: { duration: 0.6, ease: 'easeOut' },
+      },
+    }),
+    []
+  );
+
+  return { containerVariants, itemVariants };
+}
