+++ components/Cards/MeetupCard/index.ts
@@ -1,0 +1,2 @@
+export { default } from './MeetupCard';
+export type { MeetupCardProps } from './MeetupCard';
